requirements.txt
autopep8==1.5.7
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
protobuf==3.16.0
pycodestyle==2.7.0
six==1.16.0
toml==0.10.2
Werkzeug==1.0.1
